feat: implementação de menu interativo e integração com TheAudioDB
- Implementado menu principal com opções de cadastro e consulta
- Adicionado cadastro de artista e música com relacionamento
- Implementada listagem de artistas e músicas
- Criado método de busca de música por nome do artista
- Integrada API TheAudioDB para busca de dados externos
- Ajustados models de resposta e conversão JSON
- Atualizados repositories e application.properties

+++ src/main/java/com/jhonatan/music/musicsound/repository/ArtistaRepository.java
@@ -3,6 +3,5 @@ package com.jhonatan.music.musicsound.repository;
 import com.jhonatan.music.musicsound.model.Artista;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArtistaRepository extends JpaRepository<Artista, Integer> {
-
+public interface ArtistaRepository extends JpaRepository<Artista, Long> {
 }
